added menu features
