+++ babel.config.js
@@ -13,6 +13,7 @@ module.exports = function (api) {
           safe: false,
         },
       ],
+      'react-native-reanimated/plugin',
     ],
   };
 };
