+++ package.json
@@ -11,6 +11,8 @@
     "prettier": "prettier --write **/*.js"
   },
   "dependencies": {
+    "@react-navigation/drawer": "^6.4.1",
+    "@react-navigation/native": "^6.0.10",
     "expo": "~45.0.0",
     "expo-av": "~11.2.3",
     "expo-location": "~14.2.2",
@@ -23,8 +25,12 @@
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
     "react-native-dotenv": "^3.3.1",
+    "react-native-gesture-handler": "~2.2.1",
     "react-native-maps": "0.30.2",
     "react-native-paper": "^4.12.1",
+    "react-native-reanimated": "~2.8.0",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-web": "0.17.7"
   },
   "devDependencies": {
