+++ src/Menu/SettingsMenu.js
@@ -1,0 +1,12 @@
+import React, { useEffect, useState, ReactElement, useRef } from 'react';
+import { Platform, StyleSheet, View, StatusBar, Alert, Text } from 'react-native';
+ 
+const SettingsMenu = () => {
+    return (
+        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+            <Text>Test</Text>
+        </View>
+    );
+}
+ 
+export default SettingsMenu;
